backend/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anthropic==0.23.1
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
circuitbreaker==1.4.0
click==8.1.7
cohere==5.2.4
cryptography==42.0.5
dataclasses-json==0.6.4
distro==1.9.0
exceptiongroup==1.2.0
faiss-cpu==1.8.0
fastavro==1.9.4
filelock==3.13.3
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.1
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.6
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
langchain==0.1.14
langchain-community==0.0.31
langchain-core==0.1.40
langchain-openai==0.1.1
langchain-text-splitters==0.0.1
langsmith==0.1.40
lxml==5.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
oci==2.125.2
openai==1.16.2
opensearch-py==2.5.0
oracledb==2.1.1
orjson==3.10.0
packaging==23.2
pandas==2.2.1
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.17.2
pyOpenSSL==24.1.0
pyparsing==3.1.2
pypdf==4.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.29
streamlit==1.33.0
swig==4.2.1
tenacity==8.2.3
tiktoken==0.6.0
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.18
yarl==1.9.4
Flask==3.0.3
octoai==1.0.2
flask-cors==4.0.1